requirements.txt
numpy>=1.18.0
gymnasium>=0.26
pygame>=2.2.0

torch-ac>=1.4.0
minigrid>=2.2.0
tensorboardX>=1.6
numpy>=1.3


